+++ manuskript/ui/views/outlineDelegates.py
@@ -237,11 +237,11 @@ class outlineGoalPercentageDelegate(QStyledItemDelegate):
         rect = option.rect.adjusted(margin, margin, -margin, -margin)
 
         # Move
-        rect.translate(level * rect.width() / 10, 0)
-        rect.setWidth(rect.width() - level * rect.width() / 10)
+        rect.translate(int(level * rect.width() / 10), 0)
+        rect.setWidth(int(rect.width() - level * rect.width() / 10))
 
-        rect.setHeight(height)
-        rect.setTop(option.rect.top() + (option.rect.height() - height) / 2)
+        rect.setHeight(int(height))
+        rect.setTop(int(option.rect.top() + (option.rect.height() - height) / 2))
 
         drawProgress(painter, rect, p)  # from functions
